backend/pyproject.toml
[project]
name = "gitlab-user-reports"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.1",
    "fastapi>=0.121.0",
    "jinja2>=3.1.6",
    "plotly>=6.3.1",
    "pydantic-ai>=1.10.0",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.15.3",
    "pyjwt>=2.9.0",
    "python-gitlab>=7.0.0",
    "uvicorn>=0.38.0",
    "weasyprint>=66.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
